# server/docs/swagger.yaml
definitions:
...
  model.CustomerService:
    properties:
      created_at:
        type: string
      des:
        type: string
      duration:
        type: integer
      goods_id:
        description: |-
          OriginalAmount string `json:"original_amount"  gorm:"comment:原始金额"`
          商品参数-基础参数
        type: integer
      goods_type:
        type: string
      id:
        type: integer
      is_renew:
        description: 续费参数
        type: boolean
      node_connector:
        type: integer
      node_speed_limit:
        type: integer
      price:
        type: string
      renewal_amount:
        type: string
      service_end_at:
        type: string
      service_start_at:
        type: string
      service_status:
        type: boolean
      sub_status:
        type: boolean
      sub_uuid:
        type: string
      subject:
        type: string
      total_bandwidth:
        description: 商品参数-订阅
        type: integer
      traffic_reset_day:
        type: integer
      updated_at:
        type: string
      used_down:
        type: integer
      used_up:
        type: integer
      user_id:
        description: 关联用户
        type: integer
      user_name:
        type: string
    type: object
...
  model.EpayResultResponse:
    properties:
      money:
        description: 商品金额
        type: string
      name:
        description: 商品名称
        type: string
      out_trade_no:
        description: 商户订单号
        type: string
      param:
        description: 业务扩展参数
        type: string
      pid:
        description: 商户ID
        type: integer
      sign:
        description: 签名字符串，所有参数按照参数名ASCII码从小到大排序（a-z），sign、sign_type、和空值不参与签名！sign
          = md5 ( a=b&c=d&e=f + KEY ) （注意：+ 为各语言的拼接符，不是字符！），md5结果为小写。
        type: string
      sign_type:
        description: 签名类型
        type: string
      trade_no:
        description: 易支付订单号
        type: string
      trade_status:
        description: 支付状态
        type: string
      type:
        description: 支付方式
        type: string
    type: object
  model.FieldParamsItem:
    properties:
      condition:
        description: = > < <> like
        type: string
      condition_value:
        type: string
      field:
        type: string
      field_type:
        type: string
      operator:
        description: AND OR
        type: string
    type: object
...
  model.Pagination:
    properties:
      order_by:
        type: string
      page_num:
        type: integer
      page_size:
        type: integer
    type: object
...
  model.QueryParams:
    properties:
      field_params_list:
        items:
          $ref: '#/definitions/model.FieldParamsItem'
        type: array
      pagination:
        allOf:
        - $ref: '#/definitions/model.Pagination'
        description: 分页参数
      table_name:
        type: string
    required:
    - table_name
    type: object
...
  model.Ticket:
    properties:
      created_at:
        type: string
      details:
        type: string
      id:
        type: integer
      status:
        description: TicketProcessing TicketClosed
        type: string
      ticket_message:
        description: 一对多关联
        items:
          $ref: '#/definitions/model.TicketMessage'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.TicketMessage:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      message:
        type: string
      ticket_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.ResponseStruct:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact:
    url: https://github.com/shuangyu233/AirGo_Modify
  description: AirGo前后分离,多用户代理面板
  license:
    name: GPL v3.0
    url: https://github.com/shuangyu233/AirGo_Modify/blob/main/LICENSE
  title: AirGo
paths:
...
  /api/customer/finance/getBalanceStatementList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取余额流水
      tags:
      - '[customer api] finance'
  /api/customer/finance/getCommissionStatementList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取佣金流水
      tags:
      - '[customer api] finance'
  /api/customer/finance/getCommissionSummary:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取佣金统计
      tags:
      - '[customer api] finance'
  /api/customer/finance/getInvitationUserList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取邀请人数
      tags:
      - '[customer api] finance'
  /api/customer/finance/withdrawToBalance:
    get:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 提现
      tags:
      - '[customer api] finance'
...
  /api/customer/ticket/firstTicket:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Ticket'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取工单
      tags:
      - '[customer api] ticket'
  /api/customer/ticket/getUserTicketList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.QueryParams'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取工单列表
      tags:
      - '[customer api] ticket'
  /api/customer/ticket/newTicket:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Ticket'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 新建工单
      tags:
      - '[customer api] ticket'
  /api/customer/ticket/sendTicketMessage:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.TicketMessage'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 发送工单消息
      tags:
      - '[customer api] ticket'
  /api/customer/ticket/updateUserTicket:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.Ticket'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 更新工单
      tags:
      - '[customer api] ticket'
  /api/customer/traffic/getSubTrafficList:
    post:
      parameters:
      - description: Bearer 用户token
        in: header
        name: Authorization
        type: string
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.CustomerService'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            $ref: '#/definitions/response.ResponseStruct'
      summary: 获取订阅流量记录
      tags:
      - '[customer api] traffic'
...
  /api/public/shop/epayNotify:
    get:
      parameters:
      - description: 参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.EpayResultResponse'
      produces:
      - application/json
      responses:
        "200":
          description: 请求成功；正常：业务代码 code=0；错误：业务代码code=1
          schema:
            type: string
        "400":
          description: 请求错误
      summary: 易支付异步回调
      tags:
      - '[public api] shop'
...
swagger: "2.0"